separate user and group permission specifications
